+++ src/Transport/NullTransport.php
@@ -10,6 +10,8 @@ use Sentry\Event;
  * This transport fakes the sending of events by just ignoring them.
  *
  * @author Stefano Arlandini <sarlandini@alice.it>
+ *
+ * @final since 2.3
  */
 class NullTransport implements TransportInterface
 {